backend/requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
authentication==1.1.0
bleach==6.1.0
blinker==1.7.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.5
defusedxml==0.7.1
dj-rest-auth==5.0.2
Django==5.0.2
django-allauth==0.61.1
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
drf-jwt==1.19.2
drf_api_logger==1.1.15
Flask==3.0.2
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
oauthlib==3.2.2
passlib==1.7.4
pillow==10.2.0
pycparser==2.21
pycryptodome==3.20.0
PyJWT==1.7.1
python3-openid==3.2.0
pytz==2024.1
redis==5.0.2
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.4
urllib3==2.2.1
webencodings==0.5.1
Werkzeug==3.0.1
whitenoise==6.6.0
WTForms==3.1.2
